refactor: createdAt, updatedAt 모든 도메인 공통으로 들어가여 하나의 클래스로 분리
