+++ src/main/java/nextstep/DateDomain.java
@@ -1,0 +1,27 @@
+package nextstep;
+
+import java.time.LocalDateTime;
+
+public class DateDomain {
+
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
+
+    public DateDomain() {
+        this.createdAt = LocalDateTime.now();
+        this.updatedAt = LocalDateTime.now();
+    }
+
+    public DateDomain(LocalDateTime createdAt, LocalDateTime updatedAt) {
+        this.createdAt = createdAt;
+        this.updatedAt = updatedAt;
+    }
+
+    public LocalDateTime getCreatedAt() {
+        return createdAt;
+    }
+
+    public LocalDateTime getUpdatedAt() {
+        return updatedAt;
+    }
+}
